+++ apps/api/src/lib/types/publicMiddleware.ts
@@ -1,10 +1,8 @@
-import { Octokit } from "@floe/utils";
 import { NextApiRequest, NextApiResponse, NextApiHandler } from "next";
 import { Middleware } from "next-api-middleware";
 
 export type NextApiRequestExtension = NextApiRequest & {
   keyId: string;
-  octokit: Octokit;
 };
 
 export type NextApiResponseExtension = NextApiResponse;
